ml/requirements.txt
google-generativeai==0.7.2
python-dotenv==1.0.0
pandas==2.2.0
